fix: request接口jsb和纯js实现转换
